refactor: factored out Cursor
